v.1.2.9 PID sistem güncellendi.

+++ Views/UserControls/ESC.xaml.cs
@@ -17,7 +17,7 @@ namespace Advanced_Dynotis_Software.Views.UserControls
             if (sender is Thumb thumb && thumb.TemplatedParent is Slider slider && DataContext is ESCParametersViewModel viewModel && viewModel.ESCStatus)
             {
                 double newValue = slider.Value + (e.HorizontalChange / slider.ActualWidth) * (slider.Maximum - slider.Minimum);
-                viewModel.ESCValue = Math.Round(Math.Max(slider.Minimum, Math.Min(slider.Maximum, newValue)), 0);
+                viewModel.ESCValue = (int)Math.Round(Math.Max(slider.Minimum, Math.Min(slider.Maximum, newValue)), 0);
             }
         }
     }
